fix for required_forms migration nil case

+++ db/migrate/20180524135959_remove_required_forms_tag.rb
@@ -5,9 +5,11 @@ class RemoveRequiredFormsTag < ActiveRecord::Migration[5.2]
 
   def change
     tag = Tag.find_by_name('required forms')
-    taggings = Tagging.where(tag_id: tag.id)
+    if tag
+      taggings = Tagging.where(tag_id: tag.id)
 
-    tag.destroy
-    taggings.destroy_all
+      tag.destroy
+      taggings.destroy_all
+    end
   end
 end
